requirements.txt
fastapi
pymongo
faiss-cpu
httpx
uvicorn
python-dotenv
python-multipart
Pillow